Tidyup the Freyr-Screen
