+++ freyr-screen/main.py
@@ -1,6 +1,5 @@
 import gc
 import ntptime
-from machine import WDT
 from network import STA_IF, WLAN
 from utime import sleep
 
@@ -8,7 +7,7 @@ from config import password, ssid
 from freyr_screen import FreyrScreen
 
 wlan = WLAN(STA_IF)
-watchdog = WDT(timeout=8000)  # 8 Seconds
+# watchdog = WDT(timeout=8388)  # noqa: ERA001
 
 
 def connect_to_wifi() -> None:
@@ -34,7 +33,7 @@ def sleep_min(value: int = 1) -> None:
     for _ in range(value):
         for _ in range(12):
             gc.collect()
-            watchdog.feed()
+            # watchdog.feed()  # noqa: ERA001
             sleep(5)
